Feature: Refactor module with sub test dir and module naming convention
