+++ usermanagement/backend/adapter/tests/UserAuthentication_test.go
@@ -1,6 +1,7 @@
-package adapter
+package tests
 
 import (
+	adapter "github.com/D4ykoo/usermanagement/backend"
 	"regexp"
 	"testing"
 )
@@ -11,7 +12,7 @@ var jwtSecret = "your-fav-secret"
 // Checking for a valid return value.
 func TestJWTBuilding(t *testing.T) {
 	username := "test"
-	got, err := createJWT(username, jwtSecret, true)
+	got, err := adapter.CreateJWT(username, jwtSecret, true)
 
 	// generated with: https://jwt.io
 	want := "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjY4NjYyMDgwMCwidXNlcm5hbWUiOiJ0ZXN0In0.CbmybuROnf_3ClsxXiYiTbK26Dc0e2zSwMeCZZz4guszI-q8LL6HO42HJTeAjQ0gDFRmL4PikQoP8QzdPC03yw"
@@ -23,7 +24,7 @@ func TestJWTBuilding(t *testing.T) {
 
 func TestJWTValidating(t *testing.T) {
 	token := "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjY4NjYyMDgwMCwidXNlcm5hbWUiOiJ0ZXN0In0.CbmybuROnf_3ClsxXiYiTbK26Dc0e2zSwMeCZZz4guszI-q8LL6HO42HJTeAjQ0gDFRmL4PikQoP8QzdPC03yw"
-	valid, err, claims := validateJWT(token, jwtSecret)
+	valid, err, claims := adapter.ValidateJWT(token, jwtSecret)
 
 	if !valid || err != nil {
 		t.Fatalf(`validateJWT(%q, %q) = %t, %q, %q, is not true, nil, claims`, token, jwtSecret, valid, err, claims)
